try_catch

+++ app/Http/Controllers/CommentController.php
@@ -11,38 +11,57 @@ class CommentController extends Controller
 {
     public function store(Request $request)
     {
-        $chat_id = $request -> chat_id;
-        $comment = new Comment;
-        $comment -> user_id = Auth::User()->id;
-        $comment -> chat_id = $chat_id;
-        $comment -> comment  = $request -> comment_text;
-        $comment -> save();
+        try{
+            $chat_id = $request -> chat_id;
+            $comment = new Comment;
+            $comment -> user_id = Auth::User()->id;
+            $comment -> chat_id = $chat_id;
+            $comment -> comment  = $request -> comment_text;
+            $comment -> save();
+        }catch(\Exception $e){
+            $e->getMessage();
+        }
 
         return redirect()->route('comment.show',$chat_id);
     }
 
     public function show(Chat $chat)
     {
-        $chat = chat::where('id',$chat->id)->first();
-        $comments = Comment::where('chat_id',$chat->id)->get();
+        try{
+            $chat = chat::where('id',$chat->id)->first();
+            $comments = Comment::where('chat_id',$chat->id)
+            ->orderBy('updated_at', 'desc')->get();
+        }catch(\Exception $e){
+            $e->getMessage();
+        }
 
         return view('comment.show',compact('chat','comments'));
     }
 
     public function edit(Request $request)
     {
-        $chat = $request -> chat_id;
-        $comment =Comment::find($request -> comment_id);
-        $comment -> comment  =  $request -> comment_text;
-        $comment -> save();
+        try{
+            $chat = $request -> chat_id;
+            $comment =Comment::find($request -> comment_id);
+            $comment -> comment  =  $request -> comment_text;
+            $comment -> save();
+        }catch(\Exception $e){
+            $e->getMessage();
+        }
+
         return redirect()->route('comment.show',$chat);
     }
 
     public function destroy(chat $chat, Comment $comment)
     {
-        // $this->authorize('destroy',$comment);
-        $comment = Comment::find($comment -> id);
-        $comment->delete();
+        try{
+            $this->authorize('destroy',$comment);
+            $comment = Comment::find($comment -> id);
+            $comment->delete();
+        }catch(\Exception $e){
+            $e->getMessage();
+        }
+
         return redirect()->route('comment.show',$chat->id);
     }
 }
